backend/agentcore/agentcore-metering/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agentcore-metering"
version = "0.1.0"
description = "Unified observability for LLM/text, image, and video calls (Django adapter)"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "django>=4.2",
    "djangorestframework>=3.14",
    "drf-spectacular>=0.27.0",
    "litellm>=1.0.0",
    "json-repair>=0.25.0",
    "pyyaml>=6.0",
    "celery>=5.0",
    "agentcore-task",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-django",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["agentcore_metering*"]

[tool.setuptools.package-data]
"agentcore_metering.adapters.django.llm_static" = ["*.yaml", "providers/*.yaml"]

[tool.pytest.ini_options]
django_settings = "tests.settings"
python_files = ["test_*.py"]
addopts = "-v"
testpaths = ["tests"]
markers = [
    "unit: unit tests (no HTTP)",
    "api: API/integration tests (HTTP endpoints)",
]
